dom/tests/mochitest/bugs/test_sizetocontent_clamp.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=764240
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 764240</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=764240">Mozilla Bug 764240</a>
<p id="display"></p>
<div id="content">
  <button onclick="test();">run test</button>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 764240 **/

SimpleTest.waitForExplicitFinish();

// Error margin allowed for the window's size.
// The window size being set by the platform, we can't guarantee it will be respected.
// In platforms other than Windows, 5 is a good epsilon. On windows, it seems
// that most of the time, the width is 14/15 higher than expected so we need
// to have an epsilon that matches this difference.
// Actually, on platforms other than Windows, epsilon of 0 might work.
// Only innerWidth on Windows has a 14/15 pixels difference to what is
// requested.
var epsilon =  navigator.platform.indexOf("Win") == -1 ? 5 : 20;

function test() {
  var w = window.open('data:text/html,null', null, 'width=300,height=300');
  var nbResize = 0;

  SimpleTest.waitForFocus(function() {
    w.onresize = function() {
      nbResize++;

      if (nbResize == 1) {
        return;
      }

      ok(w.innerWidth + epsilon >= 100 && w.innerWidth - epsilon <= 100,
         "innerWidth should be around 100");
      ok(w.innerHeight + epsilon >= 100 && w.innerHeight - epsilon <= 100,
         "innerHeight should be around 100");

      // It's not clear why 2 events are coming...
      is(nbResize, 2, "We should get 2 events.");

      w.close();

      SimpleTest.waitForFocus(function() {
        SimpleTest.finish();
      });
    };
    w.sizeToContent();
  }, w);
}

SimpleTest.waitForFocus(function() {
  synthesizeMouseAtCenter(document.getElementsByTagName('button')[0], {});
});

</script>
</pre>
</body>
</html>
